Obtain viscosity(height) assuming speed's stable
In `theory.tex` I described what we can do if viscosity depends on
depth. This turned out to be true: the closer to the bottom, the
greater the viscosity.

The script used.
- new file:   processing2.py

The script produces a plot and two csvs:
- new file:   figures/viscosity_at_height.pdf
- new file:   viscosity_at_height.csv
- new file:   viscosity_at_height_averaged.csv

Old file used to get familiar with `cv` library. Useless, so deleted.
- deleted:    trying_object_detection.py

Moved old plot to `figures/`
- renamed:    fancy_plot.pdf -> figures/fancy_plot.pdf